+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased] - TBD
 
+### Changed
+
+* moved `GETPagination`, `POSTPagination`, `GETTokenPagination` and `POSTTokenPagination` to `stac_fastapi.extensions.core.pagination.request` submodule [#717](https://github.com/stac-utils/stac-fastapi/pull/717)
+
 ## [3.0.0a4] - 2024-06-27
 
 ### Fixed